'added_sentiment_analysis'

+++ twitter.py
@@ -9,8 +9,9 @@ import re
 from nltk.corpus import stopwords
 import random
 from datetime import datetime
-
-#from scheduler import val
+from nltk.sentiment.vader import SentimentIntensityAnalyzer
+import os.path
+from os import path
 
 #from MovieCloud import settings
 #from django.core.management import setup_environ
@@ -18,21 +19,63 @@ from datetime import datetime
 import django
 import os
 
+
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "MovieCloud.settings")
 django.setup()
 
 from movie.models import *
 
+
+def Sentiment_analysis(temp , var):
+    
+    sid = SentimentIntensityAnalyzer()
+    
+    #temp = '''Ponle Ilagyna Upayogapaduthundhi 🤙🏽#Manmadhudu2 https://t.co/QgfO1MKBGM '''
+    #temp = "Surya is best"
+
+    #temp = "<<UNK>>".join(word_cloud.Tweet.values)
+    #print(temp)
+        
+    #print("***********************************************")
+    temp = re.sub("RT ","",temp)
+    temp = re.sub("(#[\d\w]+)|@[\d\w]+","",temp)
+    temp = re.sub("https.*\s","",temp)
+
+    temp = re.sub('[!"#$%&\'()*+,-/:;<=>?@[\\]^_`{|}~]',"",temp)
+                        
+    temp = re.sub('[^a-zA-Z0-9\s.]+',"",temp)
+    #print("****")
+    #print(temp)
+    
+        #temp = temp.replace("\n","!")
+    score = sid.polarity_scores(temp)
+    
+    #print(score)
+    
+    return score[var]
+
+
+
+
 stopwords = set(stopwords.words('english'))
 def pre_processing(sentence):
+    
+    sentence = re.sub("RT ","",sentence)
+    sentence = re.sub("(#[\d\w]+)|@[\d\w]+","",sentence)
+    sentence = re.sub("https.*\s","",sentence)
+    #sentence = re.sub('[^a-zA-Z0-9\s.]+',"$",sentence)
+
+    
     lemmatizer = WordNetLemmatizer()
     sentence = sentence.lower()
     #sentence = re.sub("\d+"," ",sentence)
     sentence = re.sub('[!"#$%&\'()*+,-./:;<=>?@[\\]^_`{|}~]'," ",sentence)
+                        
+    
     stem = ""
-    Actual_Words = ["JJ","JJR","JJS",\
-                    "RB","RBR","RBS"\
-                    "UH",]
+    Actual_Words = ["JJ","JJR","JJS",]#\
+                    #"RB","RBR","RBS"\
+                    #"UH",]
                     #"VBG","VBP"]
     Total_Words=[]
 
@@ -41,9 +84,9 @@ def pre_processing(sentence):
     #random.shuffle(Total_Words)
     
 
-    print("----",)
-    print(nltk.pos_tag(nltk.word_tokenize(sentence)))
-    print("***************")
+    #print("----",)
+    #print(nltk.pos_tag(nltk.word_tokenize(sentence)))
+    #print("***************")
     for word in nltk.pos_tag(nltk.word_tokenize(sentence)):
         if word[0] not in stopwords:
             #print("-----",nltk.pos_tag(word))
@@ -52,10 +95,10 @@ def pre_processing(sentence):
     for i in stem.split():
         Total_Words.append(nltk.pos_tag([i])[0])
         
-    print(Total_Words)
+    #print(Total_Words)
 
     Final_Words = [word[0] for word in Total_Words if word[1] in Actual_Words]
-    print([item for item in Total_Words if item[1]=="JJ"])
+    #print([item for item in Total_Words if item[1]=="JJ"])
 
     return " ".join(Final_Words)
 
@@ -158,24 +201,45 @@ class Twitter():
 
 #print(item.get_token())
 def func_twitter():
-    metro_polices = ["#Manmadhudu2"]
+    metro_polices = ["#Saaho"]
     from pprint import pprint
     #[pprint(item["text"]) for item in police["statuses"]]
+    
+
     public_requests_police = pd.DataFrame()
     for index,search_query in enumerate(metro_polices):
         item = Twitter()
         item.get_token()
-        #created object
-        parameter = {
-        'q': search_query,
-        'result_type': "recent",
-        'tweet_mode' : "extended",
-        'count': 100,
-        #'include_entities':1,
-        }
+        
+        if(not path.exists('media/Excel/'+str(search_query)+'.xlsx')):
+                           print("Nothing")
+                           
+                           #created object
+                           parameter = {
+                           'q': search_query,
+                           'result_type': "recent",
+                           'tweet_mode' : "extended",
+                           'count': 100,
+                            #'include_entities':1,
+                           }
+        else:
+           reading = pd.read_excel("media/Excel/"+search_query+".xlsx",usecols = [5],index = False)
+           
+           parameter = {
+           'max_id' : str(reading.iloc[-1,0]),
+           'q': search_query,
+           'result_type': "recent",
+           'tweet_mode' : "extended",
+           'count': 100,
+            #'include_entities':1,
+           }
+           print("max____ID: ",parameter['max_id'])
+
+            
+                           
         rows_count = 0
         rows = []
-        while(rows_count<=1000):
+        while(rows_count<=500):
             #total_count+=1
             print("search_params : ",parameter)
     
@@ -209,18 +273,20 @@ def func_twitter():
                     continue
                 
                 rows_count = rows_count+1
-                rows.append([user["user"]["name"],user["user"]["screen_name"],user["user"]["location"],user["user"]["description"]\
+                
+                
+                rows.append([user["user"]["name"],user["user"]["screen_name"],user["user"]["location"],user["user"]["description"],user["id_str"]\
                 ,user["full_text"],user["in_reply_to_screen_name"],str(get_mentioned_urls(user)),user["lang"],"https://twitter.com/statuses/"+user["id_str"],is_retweeted,who_retweeted\
                 ,str(get_hashtags(user)),str(get_user_mentions(user)), get_date_format(user["created_at"]), get_time_format(user["created_at"]) , user["user"]["verified"]  ])
             if "next_results" not in police["search_metadata"]:
                 print("main_sdfjaksdljfaklsd ---->>>",police)
                 break
             #print(rows)
-            print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
-            print("????????????????",police["search_metadata"])
+            #print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
+            #print("????????????????",police["search_metadata"])
             next = police["search_metadata"]["next_results"]
-            print(next)
-            print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!",)
+            #print(next)
+            #print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!",)
     
             new_params = {item.split('=')[0]:item.split('=')[1] for item in next[1:].split('&')}
             #parameter[""]= police["search_metadata"]["next_results"]
@@ -231,12 +297,25 @@ def func_twitter():
             parameter["q"] = search_query
             #print("^^^^^^^",total_count,"^^^^^^")
         #print(rows)
-        tweets = pd.DataFrame(rows,columns = ["User_Name", "Twitter_Name","Location", "Bio", "Tweet","Replied_to","media","Language","Tweet_Link",\
+        tweets = pd.DataFrame(rows,columns = ["User_Name", "Twitter_Name","Location", "Bio","Tweet_id", "Tweet","Replied_to","media","Language","Tweet_Link",\
         "is_Retweeted","Original_Tweet_by","Hashtags","User_Mentions","tweet_date","tweet_time","Verified"])
         #print(rows)
         #print("these are tweets:::",tweets)
         public_requests_police = public_requests_police.append(tweets,ignore_index = True)
-        tweets.to_excel("media/Excel/"+search_query+".xlsx")
+        
+        tweets['Positive'] = tweets['Tweet'].apply((lambda x:Sentiment_analysis(x,'pos')))
+        tweets['Negitive'] = tweets['Tweet'].apply((lambda x:Sentiment_analysis(x,'neg')))
+        if(path.exists('media/Excel/'+str(search_query)+'.xlsx')):
+            new_updated_df = pd.read_excel("media/Excel/"+search_query+".xlsx")
+            new_updated_df.append(tweets,ignore_index = True)
+            new_updated_df.to_excel("media/Excel/"+search_query+".xlsx")
+
+            
+        else:
+            tweets.to_excel("media/Excel/"+search_query+".xlsx")
+        
+        
+        
         #pprint(police)
     #public_requests_police.to_excel("public_requests_police.xlsx")
     
@@ -244,11 +323,33 @@ def func_twitter():
     
     for i in metro_polices:
         word_cloud = pd.read_excel("media/Excel/"+i+".xlsx",columns = ["Tweet"])
+
+
+        #temp = "".join(word_cloud.Tweet.values)
+        
         temp = word_cloud.Tweet.to_string()
+        '''
+        print(temp)
+        
+        print("***********************************************")
         temp = re.sub("RT ","",temp)
         temp = re.sub("(#[\d\w]+)|@[\d\w]+","",temp)
         temp = re.sub("https.*\s","",temp)
+
+        temp = re.sub('[!"#$%&\'()*+,-/:;<=>?@[\\]^_`{|}~]',"",temp)
+                        
+        temp = re.sub('[^a-zA-Z0-9\s.]+',"$",temp)
+        
+        #temp = temp.replace("\n","!")
+
+        print(temp)
+        
+        print(type(temp))
+        print(len(temp))
+        '''
+
         processed_temp = pre_processing(temp)
+
         #print(processed_temp)
         wordcloud = WordCloud(width = 800, height = 800,
                     background_color ='white',
@@ -268,8 +369,8 @@ def func_twitter():
         if created:
             instance = Movie.objects.get(movie_name = i)
             instance.delete()
-            Movie.objects.create(movie_name=i , poster = "media/image/"+i+".jpg", 
-                                 word_cloud = "../media/image/"+i+".jpg" , 
+            Movie.objects.create(movie_name=i , poster = "image/"+i+".jpg", 
+                                 word_cloud = "image/"+i+".jpg" , 
                                  last_updated =datetime.now() ).save()
   
             print("8888888888888888888888888888888888888888888888888888888888888888")
@@ -278,10 +379,10 @@ def func_twitter():
         else:
             obj = Movie.objects.get(movie_name=i)
             obj.last_updated = datetime.now()
+            obj.poster = "image/"+i+".jpg"
             obj.save()
             print("8888888888888888888888888888888888888888888888888888888888888888")
             print("inside else")
-            
             # person just refers to the existing one
         
             
@@ -290,6 +391,8 @@ def func_twitter():
         
         for i in Movie.objects.all():
             print(i.movie_name,i.last_updated,i.word_cloud)
-
-#func_twitter()        
         
+        max_id = parameter['max_id']
+
+func_twitter()        
+#Sentiment_analysis()
